Move DisplayStyleSwitcher to editor-components
