+++ plugins/woocommerce/changelog/update-display-style-switch-to-editor-components
@@ -1,0 +1,5 @@
+Significance: patch
+Type: tweak
+Comment: Move DisplayStyleSwitcher to editor-components
+
+
