+++ plugins/woocommerce/client/blocks/assets/js/blocks/product-filters/constants.ts
@@ -9,11 +9,3 @@ export const EXCLUDED_BLOCKS = [
 	'woocommerce/add-to-cart-with-options',
 	'core/query',
 ];
-
-/**
- * Block types that declare a product filter or variation-selector parent in `ancestor`
- * but are not interchangeable display styles for the style toggle (chips vs dropdown).
- */
-export const DISPLAY_STYLE_SWITCHER_EXCLUDED_BLOCK_NAMES: string[] = [
-	'woocommerce/add-to-cart-with-options-variation-selector-attribute-name',
-];
