+++ plugins/woocommerce/client/blocks/assets/js/editor-components/display-style-switcher/constants.ts
@@ -1,0 +1,7 @@
+/**
+ * Block types that declare a product filter or variation-selector parent in `ancestor`
+ * but are not interchangeable display styles for the style toggle (chips vs dropdown).
+ */
+export const DISPLAY_STYLE_SWITCHER_EXCLUDED_BLOCK_NAMES: string[] = [
+	'woocommerce/add-to-cart-with-options-variation-selector-attribute-name',
+];
